requirements.txt
PyYAML>=6.0
jsonschema>=4.0.0
requests>=2.25.0
click>=8.0.0
pathlib2>=2.3.0
python-dateutil>=2.8.0
semver>=2.13.0 
pandas>=2.0.0 